feat: SAP-level master data architecture + E2E harmonization
AccountMaster (számlatükör):
- New account_master table with 33 accounts (natural key: code)
- FK constraints on budget_lines, accounting_entries, purchase_orders, invoices, templates
- Accounts API: GET /list, /tree, /{code}, POST, PATCH
- Account code validation on PO create and budget line create
- Auto-fill account_name and pnl_category from master

Partner Master enrichment:
- 12 new fields: payment_terms, city, iban, swift, contact, is_verified etc.
- Partner detail panel on frontend with all enrichment data
- Partner dropdown on PO creation (auto-fills supplier name, tax_id, accounting_code)
- Partner-based PO matching in reconciliation service

Float → Numeric migration:
- All monetary amounts: Numeric(18,2), rates: Numeric(5,2), scores: Numeric(5,4)
- Decimal→float conversion in all service layers for JSON serialization

Department hierarchy:
- Igazgatóság (MGMT) as root, all depts under it
- 18 users across 8 departments with full position hierarchy
- CEO → CFO → dept heads → workers approval chain
- Budget lines seeded for FIN and LOG departments

Role-based PO visibility:
- admin/cfo/accountant see all POs
- dept_head/clerk/reviewer see only own department
- Frontend: department filter locked for non-admin users

Invoice unique index:
- Partial unique on (partner_id, invoice_number) WHERE NOT duplicate

Compact NewOrderPage:
- 2-column layout: form left, budget panel right (sticky)
- Collapsible active PO list
- All fields fit on one screen

+++ backend/common/common/models/accounting_entry.py
@@ -1,7 +1,8 @@
 import uuid
 import enum
 from datetime import datetime
-from sqlalchemy import String, Float, DateTime, ForeignKey, Enum as SAEnum
+from decimal import Decimal
+from sqlalchemy import String, DateTime, ForeignKey, Numeric, Enum as SAEnum
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from common.database import Base
 
@@ -19,7 +20,7 @@ class AccountingEntry(Base):
     purchase_order_id: Mapped[str | None] = mapped_column(String(36), ForeignKey("purchase_orders.id"), nullable=True)
     account_code: Mapped[str] = mapped_column(String(20), nullable=False)
     department_id: Mapped[str] = mapped_column(String(36), ForeignKey("departments.id"), nullable=False)
-    amount: Mapped[float] = mapped_column(Float, nullable=False)
+    amount: Mapped[Decimal] = mapped_column(Numeric(18, 2), nullable=False)
     currency: Mapped[str] = mapped_column(String(3), default="HUF", nullable=False)
     period: Mapped[str] = mapped_column(String(7), nullable=False)  # YYYY-MM
     entry_type: Mapped[EntryType] = mapped_column(SAEnum(EntryType), nullable=False)
