+++ backend/common/common/models/budget_line.py
@@ -1,7 +1,8 @@
 import uuid
 import enum
 from datetime import datetime
-from sqlalchemy import String, Float, Integer, DateTime, ForeignKey, Enum as SAEnum
+from decimal import Decimal
+from sqlalchemy import String, Integer, DateTime, ForeignKey, Numeric, Enum as SAEnum
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from common.database import Base
 
@@ -34,7 +35,7 @@ class BudgetLine(Base):
     account_code: Mapped[str] = mapped_column(String(20), nullable=False)
     account_name: Mapped[str] = mapped_column(String(255), nullable=False)
     period: Mapped[str] = mapped_column(String(7), nullable=False)  # YYYY-MM
-    planned_amount: Mapped[float] = mapped_column(Float, nullable=False)
+    planned_amount: Mapped[Decimal] = mapped_column(Numeric(18, 2), nullable=False)
     currency: Mapped[str] = mapped_column(String(3), default="HUF", nullable=False)
     status: Mapped[BudgetStatus] = mapped_column(SAEnum(BudgetStatus), default=BudgetStatus.draft, nullable=False)
     pnl_category: Mapped[str] = mapped_column(String(20), default="opex", nullable=False)
