+++ backend/common/common/models/purchase_order_line.py
@@ -1,5 +1,6 @@
 import uuid
-from sqlalchemy import String, Float, Integer, ForeignKey
+from decimal import Decimal
+from sqlalchemy import String, Integer, ForeignKey, Numeric
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from common.database import Base
 
@@ -10,9 +11,9 @@ class PurchaseOrderLine(Base):
     id: Mapped[str] = mapped_column(String(36), primary_key=True, default=lambda: str(uuid.uuid4()))
     purchase_order_id: Mapped[str] = mapped_column(String(36), ForeignKey("purchase_orders.id", ondelete="CASCADE"), nullable=False)
     description: Mapped[str] = mapped_column(String(500), nullable=False)
-    quantity: Mapped[float] = mapped_column(Float, nullable=False)
-    unit_price: Mapped[float] = mapped_column(Float, nullable=False)
-    net_amount: Mapped[float] = mapped_column(Float, nullable=False)
+    quantity: Mapped[Decimal] = mapped_column(Numeric(18, 2), nullable=False)
+    unit_price: Mapped[Decimal] = mapped_column(Numeric(18, 2), nullable=False)
+    net_amount: Mapped[Decimal] = mapped_column(Numeric(18, 2), nullable=False)
     sort_order: Mapped[int] = mapped_column(Integer, default=0, nullable=False)
 
     purchase_order = relationship("PurchaseOrder", back_populates="lines")
